+++ extensions/vanguard-backup/CHANGELOG.md
@@ -1,3 +1,7 @@
 # Vanguard Backup Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- Manage **Backup Destinations**
+- Manage **Backup Tasks**
+- Manage **Remote Servers**